requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-environ>=0.11
psycopg2-binary>=2.9
gunicorn>=22.0
drf-spectacular>=0.27